requirements.txt
numpy
pandas
faiss-cpu
sentence-transformers

